infrastructure/cdk/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
aws-cdk.aws-elasticloadbalancingv2-alpha
aws-cdk.aws-servicediscovery
aws-cdk.aws-certificatemanager
aws-cdk.aws-route53
aws-cdk.aws-route53-targets